+++ samples/Logfella.Samples.MVC/Startup.cs
@@ -27,7 +27,7 @@ namespace Logfella.Samples.MVC
             // one has full control over how or which `ILogWriter` to use and to inspect an incoming
             // HttpRequest in order to dynamically configure the instance (e.g. check the HTTP
             // headers for an X-Correlation-Id header and use that value as `requestId`):
-            app.UseRequestBasedLogWriter(
+            app.UseRequestScopedLogWriter(
                 ctx =>
                     // Note that one can use an already existing and pre-configured GoogleCloudLogWriter
                     // to extend it with additional features without having to re-create the entire
@@ -49,7 +49,8 @@ namespace Logfella.Samples.MVC
                 app.UseDeveloperExceptionPage();
 
             app .UseRequestLogging(
-                    excludeHttpHeaders: HashSet.New("Authorization", "Cookie", "X-ApiKey"))
+                    options =>
+                        options.ExcludeHttpHeaders = HashSet.New("Authorization", "Cookie", "X-ApiKey"))
                 .UseRouting()
                 .UseEndpoints(
                     endpoints => { endpoints.MapControllers(); });